Switching alignment strategy for grid cells to be more consistent with other layout types
Developer tools in Chrome seems to indicate that align-self and justify-self don't work in this context, but further investigation reveals that they do, and the information in Chrome is in error. I believe this is happening because Chrome sees that the child cell is not configured with display: grid, but for *-self properties, it is the parent that matters, not the child itself.