Panel components now auto-root themselves (Issue #4)
The Root component is eliminated. Panel components now have a default width and height of 100%, which the user can override. However, in order to allow panels to nest inside other panels, width and height for element components are tagged with !important so that element sizing rules always take precedence. This system allows most panel uses to auto root themselves without any work on the part of the user, while allowing full control when a panel is a true root.